Moved SharedBufferManager module and test cases into odin_data package
